update primitive utils tests
